# perf: speed up latn number formatting