# WIP

```console
✗ benchstat a.txt b.txt                                
goos: darwin
goarch: arm64
pkg: go.expect.digital/intl
cpu: Apple M2 Max
                               │    a.txt     │                b.txt                │
                               │    sec/op    │   sec/op     vs base                │
FmtEraYearMonthDayGregorian-12   122.70n ± 1%   91.56n ± 0%  -25.38% (p=0.000 n=10)
NewDateTime/fa-IR-12              27.94n ± 0%   28.77n ± 0%   +2.97% (p=0.000 n=10)
NewDateTime/lv-LV-12              22.67n ± 0%   23.23n ± 0%   +2.51% (p=0.000 n=10)
NewDateTime/dz-BT-12              22.09n ± 1%   22.62n ± 1%   +2.42% (p=0.001 n=10)
DateTime_Format/fa-IR-12          576.0n ± 1%   600.9n ± 1%   +4.32% (p=0.000 n=10)
DateTime_Format/lv-LV-12          362.5n ± 0%   265.3n ± 1%  -26.81% (p=0.000 n=10)
DateTime_Format/dz-BT-12          378.4n ± 0%   415.8n ± 1%   +9.86% (p=0.000 n=10)
geomean                           104.4n        98.78n        -5.43%

                               │    a.txt     │                b.txt                 │
                               │     B/op     │    B/op     vs base                  │
FmtEraYearMonthDayGregorian-12   44.00 ± 0%     32.00 ± 0%  -27.27% (p=0.000 n=10)
NewDateTime/fa-IR-12             0.000 ± 0%     0.000 ± 0%        ~ (p=1.000 n=10) ¹
NewDateTime/lv-LV-12             0.000 ± 0%     0.000 ± 0%        ~ (p=1.000 n=10) ¹
NewDateTime/dz-BT-12             0.000 ± 0%     0.000 ± 0%        ~ (p=1.000 n=10) ¹
DateTime_Format/fa-IR-12         152.0 ± 0%     152.0 ± 0%        ~ (p=1.000 n=10) ¹
DateTime_Format/lv-LV-12         93.00 ± 0%     72.00 ± 0%  -22.58% (p=0.000 n=10)
DateTime_Format/dz-BT-12         152.0 ± 0%     224.0 ± 0%  +47.37% (p=0.000 n=10)
geomean                                     ²                -2.63%                ²
¹ all samples are equal
² summaries must be >0 to compute geomean

                               │    a.txt     │                b.txt                 │
                               │  allocs/op   │ allocs/op   vs base                  │
FmtEraYearMonthDayGregorian-12   4.000 ± 0%     4.000 ± 0%        ~ (p=1.000 n=10) ¹
NewDateTime/fa-IR-12             0.000 ± 0%     0.000 ± 0%        ~ (p=1.000 n=10) ¹
NewDateTime/lv-LV-12             0.000 ± 0%     0.000 ± 0%        ~ (p=1.000 n=10) ¹
NewDateTime/dz-BT-12             0.000 ± 0%     0.000 ± 0%        ~ (p=1.000 n=10) ¹
DateTime_Format/fa-IR-12         11.00 ± 0%     11.00 ± 0%        ~ (p=1.000 n=10) ¹
DateTime_Format/lv-LV-12         12.00 ± 0%     12.00 ± 0%        ~ (p=1.000 n=10) ¹
DateTime_Format/dz-BT-12         11.00 ± 0%     13.00 ± 0%  +18.18% (p=0.000 n=10)
geomean                                     ²                +2.42%              
```